封装获取验证码组件
